CMRARC-355 added features and few other minor changes
