+++ test/controllers/login_controller_test.rb
@@ -1,27 +1,10 @@
 require 'test_helper'
 
+Dir[Rails.root.join("test/**/*.rb")].each {|f| require f}
+
 # https://gist.github.com/kinopyo/1338738
 class LoginControllerTest < ActionController::TestCase
-
-  def mock_auth_hash
-
-    omniauth_hash = {'provider' => 'urs',
-                     'uid' => '12345',
-                     'info' => {
-                       'first_name' => 'chris',
-                       'last_name' => 'gokey',
-                       'email_address' => 'cgokey@sgt-inc.com',
-                     },
-                     'extra' => {'raw_info' =>
-                                   {'user_type' => 'my_user_type',
-                                    'organization' => 'NASA'
-                                   }
-                     }
-    }
-    OmniAuth.config.add_mock(:urs, omniauth_hash)
-
-
-  end
+  include OmniauthMacros
 
   describe "POST #urs" do
     before do
@@ -40,23 +23,6 @@ class LoginControllerTest < ActionController::TestCase
       end
     end
 
-    describe "access top page" do
-      it "can sign in user with oauth account" do
-        mock_auth_hash
-        visit '/'
-        page.must_have_content("Login with Earth Data Login")
-        click_link "Login"
-        page.must_have_content("Logout")
-      end
-
-      it "can handle authentication error" do
-        OmniAuth.config.mock_auth[:urs] = :invalid_credentials
-        visit '/'
-        page.must_have_content("Login with Earth Data Login")
-        click_link "Login"
-        page.must_have_content('Could not authenticate you from URS because "Invalid credentials".')
-      end
-    end
-
   end
 end
+
